requirements.txt
asgiref==3.9.1
attrs==25.4.0
colorama==0.4.6
dj-database-url==2.1.0
Django==4.2.7
django-cors-headers==4.3.1
django-filter==25.1
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.1
dnspython==2.8.0
drf-spectacular==0.29.0
email-validator==2.3.0
gunicorn==21.2.0
idna==3.10
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
psycopg2-binary==2.9.10
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-django==4.11.1
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.28.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
whitenoise==6.6.0
